requirements.txt
pypdf
langchain-openai
pinecone-client
langchain
langchain-community
st-star-rating
st-gsheets-connection
firebase-admin
pycryptodome